# Conflicts:
#	ms-products/src/main/java/com/compassuol/sp/challenge/msproducts/model/ProductModel.java
#	ms-products/src/main/java/com/compassuol/sp/challenge/msproducts/repository/ProductRepository.java

+++ ms-products/src/main/java/com/compassuol/sp/challenge/msproducts/repository/ProductRepository.java
@@ -1,9 +1,9 @@
 package com.compassuol.sp.challenge.msproducts.repository;
 
-import org.springframework.data.jpa.repository.JpaRepository;
 import com.compassuol.sp.challenge.msproducts.model.ProductModel;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductRepository extends JpaRepository<ProductModel, Integer> {
+public interface ProductRepository extends JpaRepository<ProductModel, Long> {
 }
